move sql preview impl. and utils to new  src/views folder
- rename it to sqlPreview.ts
- and delete old src/sql_ouput folder
